+++ cpp/metrics_utils.h
@@ -21,19 +21,19 @@ namespace metrics_utils {
     //    x1   ...   xn
     //  ( y1 ) ... ( yn )
     //    z1   ...   zn
-    Eigen::MatrixX3d points_idx_to_matrix(cilantro::VectorSet3d const & points, std::vector<unsigned long> const & idx);
+    Eigen::MatrixX3d points_idx_to_matrix(cilantro::VectorSet3d const & points, std::vector<int> const & idx);
 
     PointCloud aggregate_map(std::vector<PointCloud> const & pcs, std::vector<Eigen::Matrix4d> const & ts);
 
-    std::vector<unsigned long> get_radius_search_indices(KDTree const & tree,
+    std::vector<int> get_radius_search_indices(KDTree const & tree,
                                                        Eigen::Vector3d const & query, double radius);
 
     namespace metrics_algorithm{
         std::optional<double> compute_eigenvalues(cilantro::VectorSet3d const & points,
-                                                  std::vector<unsigned long> const & indices);
+                                                  std::vector<int> const & indices);
 
         std::optional<double> compute_entropy(cilantro::VectorSet3d const & points,
-                                              std::vector<unsigned long> const & indices);
+                                              std::vector<int> const & indices);
     } // namespace metrics_algorithm
 } // namespace metrics_utils
